Moved the get_cte_query function from mssql.py to base.py for using irrespetcive of dbengine
